+++ crates/base/core/src/media/visual/mod.rs
@@ -8,8 +8,8 @@
 
 #[cfg(feature = "color")]
 pub mod color;
-// #[cfg(feature = "draw")]
-// pub mod draw;
+#[cfg(feature = "draw")]
+pub mod draw;
 #[cfg(feature = "image")]
 pub mod image;
 // pub mod video;
@@ -17,7 +17,7 @@ pub mod image;
 crate::structural_mods! { // _pub_mods, _crate_internasl
     _pub_mods {
         #[cfg(feature = "color")] pub use super::color::_all::*;
-        // #[cfg(feature = "draw")] pub use super::draw::_all::*;
+        #[cfg(feature = "draw")] pub use super::draw::_all::*;
         #[cfg(feature = "image")] pub use super::image::_all::*;
         // pub use super::video::_all::*;
     }